Worked on Nav Bar
Working on navBar while learning how to control element placement with CSS. Updates the reset CSS to get rid of inherited styles on the anchor tags and list elements. Set the box sizing to border box for easier control and styled the nav bar items with border, margins and padding.